이미지 업로드 수정 from Eun
